Add servlets for models

+++ src/main/scala/com/github/yogishadixit/consensus/models/User.scala
@@ -3,4 +3,4 @@ package com.github.yogishadixit.consensus.models
 /**
   * Created by yogishadixit on 3/11/16.
   */
-case class User(id: Long, name: String, email: String)
+case class User(id: Int, name: String, email: String)
